python-service/requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.3
python-dotenv==1.0.1

# Database (SQLAlchemy + SQLite)
sqlalchemy==2.0.36
aiosqlite==0.20.0

# RAG and Vector Store
llama-index==0.12.7
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.3

# OpenAI
openai==1.57.2

# Document Processing
beautifulsoup4==4.12.3
lxml==5.3.0
readability-lxml==0.8.1
html2text==2024.2.26

# Web Scraping
requests==2.32.3
